Refactor reports classes
I've refactored the reports module to make the code slightly more readable.
